moved all user variables actions to dataActions
